Changed the varibles in the .c files to double

+++ Prac-02/C/src/CHeterodyning.c
@@ -1,7 +1,7 @@
 #include "CHeterodyning.h"
 
-extern float data [SAMPLE_COUNT];
-extern float carrier[SAMPLE_COUNT];
+extern double data [SAMPLE_COUNT];
+extern double carrier[SAMPLE_COUNT];
 
 float result [SAMPLE_COUNT];
